Tests: seperating out feature test so thsoe requiring mocking of API are in seperate file
