+++ tweetmood/tests/features/test_homepage_text.py
@@ -2,6 +2,7 @@ from django.test import LiveServerTestCase
 from selenium import webdriver
 from selenium.webdriver.firefox.options import Options
 
+
 class HelloWorldTest(LiveServerTestCase):
 
     def setUp(self):
@@ -9,7 +10,6 @@ class HelloWorldTest(LiveServerTestCase):
         options.add_argument('-headless')
         self.selenium = webdriver.Firefox(options=options)
         super(HelloWorldTest, self).setUp()
-
 
     def tearDown(self):
         self.selenium.quit()
@@ -21,22 +21,11 @@ class HelloWorldTest(LiveServerTestCase):
         body_text = selenium.find_element_by_tag_name('body').text
         assert 'How does London feel about right now?' in body_text
 
-    def test_submit_button_is_show(self):
-        selenium = self.selenium
-        selenium.get(self.live_server_url)
-        text_field = selenium.find_element_by_name('text')
-        text_field.send_keys('Test')
-        selenium.find_element_by_name('analyse').click()
-        body_text = selenium.find_element_by_tag_name('body').text
-        assert 'Test' in body_text
-
-
     def test_no_analysis_shortcut(self):
         selenium = self.selenium
         selenium.get(self.live_server_url + '/analysis')
         body_text = selenium.find_element_by_tag_name('body').text
         assert 'You did not submit to analysis' in body_text
-
 
     def test_no_result_shortcut(self):
         selenium = self.selenium
